+++ tests/test_album.py
@@ -25,6 +25,7 @@ from dateutil import tz
 import tidalapi
 from tidalapi.album import Album
 from tidalapi.exceptions import MetadataNotAvailable, ObjectNotFound
+from tidalapi.media import AudioMode, Quality
 
 from .cover import verify_image_cover, verify_video_cover
 
@@ -36,15 +37,22 @@ def test_album(session):
     assert album.type == "ALBUM"
     assert album.duration == 6712
     assert album.available
+    assert album.ad_supported_ready
+    assert album.allow_streaming
+    assert album.dj_ready
+    assert album.audio_modes == ["STEREO"]
+    assert album.audio_quality == Quality.high_lossless
     assert album.num_tracks == 22
     assert album.num_videos == 0
     assert album.num_volumes == 2
     assert album.release_date == datetime.datetime(2011, 9, 22)
+    assert album.available_release_date == datetime.datetime(2011, 9, 22)
     assert album.copyright == "Sinuz Recordings (a division of HITT bv)"
     assert album.version == "Deluxe"
     assert album.cover == "30d83a8c-1db6-439d-84b4-dbfb6f03c44c"
     assert album.video_cover is None
     assert album.explicit is False
+    assert album.premium_streaming_only is False
     assert album.universal_product_number == "3610151683488"
     assert 0 < album.popularity < 100
     assert album.artist.name == "Lasgo"
@@ -152,3 +160,37 @@ def test_album_type_single(session):
 def test_album_type_ep(session):
     album = session.album(289261563)
     assert album.type == "EP"
+
+
+def test_album_quality_atmos(session):
+    # Session should allow highest possible quality (but will fallback to highest available album quality)
+    session.audio_quality = Quality.hi_res_lossless
+    album = session.album("355472560")  # DOLBY_ATMOS
+    assert album.audio_quality == Quality.low_96k
+    assert album.audio_modes == [AudioMode.dolby_atmos]
+    assert "DOLBY_ATMOS" in album.media_metadata_tags
+
+
+def test_album_quality_max(session):
+    # Session should allow highest possible quality (but will fallback to highest available album quality)
+    session.audio_quality = Quality.high_lossless
+    album = session.album("355473696")  # MAX (LOSSLESS, 16bit/48kHz)
+    assert album.audio_quality == Quality.high_lossless
+    assert album.audio_modes == [AudioMode.stereo]
+    assert "LOSSLESS" in album.media_metadata_tags
+
+
+def test_album_quality_max_lossless(session):
+    # Session should allow highest possible quality (but will fallback to highest available album quality)
+    session.audio_quality = Quality.hi_res_lossless
+    album = session.album("355473675")  # MAX (HI_RES_LOSSLESS, 24bit/192kHz)
+    assert (
+        album.audio_quality == Quality.high_lossless
+    )  # Expected HI_RES_LOSSLESS here. TIDAL bug perhaps?
+    assert album.audio_modes == [AudioMode.stereo]
+    assert "HIRES_LOSSLESS" in album.media_metadata_tags
+
+
+def test_reset_session_quality(session):
+    # HACK: Make sure to reset audio quality to default value for remaining tests
+    session.audio_quality = Quality.default