+++ tidalapi/album.py
@@ -43,7 +43,7 @@ class Album:
     name, cover and video cover. TIDAL does this to reduce the network load.
     """
 
-    id: Optional[str] = None
+    id: Optional[int] = -1
     name: Optional[str] = None
     cover = None
     video_cover = None
@@ -67,7 +67,7 @@ class Album:
     popularity: Optional[int] = -1
     user_date_added: Optional[datetime] = None
     audio_quality: Optional[str] = ""
-    audio_modes: Optional[str] = ""
+    audio_modes: Optional[List[str]] = [""]
     media_metadata_tags: Optional[List[str]] = [""]
 
     artist: Optional["Artist"] = None